requirements.txt
streamlit
google-generativeai
pdf2image
PyPDF2

